(PC-6910): Move offerer into models file
