+++ src/pcapi/admin/install.py
@@ -18,13 +18,13 @@ from pcapi.admin.custom_views.partner_user_view import PartnerUserView
 from pcapi.admin.custom_views.pro_user_view import ProUserView
 from pcapi.admin.custom_views.user_offerer_view import UserOffererView
 from pcapi.admin.custom_views.venue_view import VenueView
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.users.models import User
 from pcapi.models import AllocinePivot
 from pcapi.models import BeneficiaryImport
 from pcapi.models import Criterion
 from pcapi.models import Feature
 from pcapi.models import Offer
-from pcapi.models import Offerer
 from pcapi.models import UserOfferer
 from pcapi.models import Venue
