+++ src/pcapi/domain/bank_information.py
@@ -1,8 +1,8 @@
 from typing import List
 
+from pcapi.core.offerers.models import Offerer
 from pcapi.domain.bank_informations.bank_informations import BankInformations
 from pcapi.domain.venue.venue_with_basic_information.venue_with_basic_information import VenueWithBasicInformation
-from pcapi.models import Offerer
 from pcapi.models.bank_information import BankInformationStatus
