+++ src/pcapi/domain/pro_offerers/paginated_offerers.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from pcapi.models import Offerer
+from pcapi.core.offerers.models import Offerer
 
 
 class PaginatedOfferers:
