+++ src/pcapi/emails/new_offerer_validation.py
@@ -1,6 +1,6 @@
 from typing import Dict
 
-from pcapi.models import Offerer
+from pcapi.core.offerers.models import Offerer
 
 
 def retrieve_data_for_new_offerer_validation_email(offerer: Offerer) -> Dict:
