+++ src/pcapi/infrastructure/repository/offerer/offerer_sql_repository.py
@@ -1,7 +1,7 @@
+from pcapi.core.offerers.models import Offerer as OffererSQLEntity
 from pcapi.domain.offerer.offerer import Offerer
 from pcapi.domain.offerer.offerer_repository import OffererRepository
 from pcapi.infrastructure.repository.offerer import offerer_domain_converter
-from pcapi.models import Offerer as OffererSQLEntity
 
 
 class OffererSQLRepository(OffererRepository):
