+++ src/pcapi/infrastructure/repository/venue/venue_with_offerer_name/venue_with_offerer_name_sql_repository.py
@@ -1,11 +1,11 @@
 from typing import List
 from typing import Optional
 
+from pcapi.core.offerers.models import Offerer
 from pcapi.domain.identifier.identifier import Identifier
 from pcapi.domain.venue.venue_with_offerer_name.venue_with_offerer_name import VenueWithOffererName
 from pcapi.domain.venue.venue_with_offerer_name.venue_with_offerer_name_repository import VenueWithOffererNameRepository
 from pcapi.infrastructure.repository.venue.venue_with_offerer_name import venue_with_offerer_name_domain_converter
-from pcapi.models import Offerer
 from pcapi.models import UserOfferer
 from pcapi.models import Venue
