+++ src/pcapi/model_creators/generic_creators.py
@@ -10,6 +10,7 @@ from shapely.geometry import Polygon
 
 from pcapi import settings
 from pcapi.core.bookings import api as bookings_api
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.offers.models import Mediation
 from pcapi.core.users.models import User
 from pcapi.domain.payments import PaymentDetails
@@ -30,7 +31,6 @@ from pcapi.models import ImportStatus
 from pcapi.models import IrisFrance
 from pcapi.models import IrisVenues
 from pcapi.models import Offer
-from pcapi.models import Offerer
 from pcapi.models import Payment
 from pcapi.models import PaymentMessage
 from pcapi.models import PaymentStatus
