+++ src/pcapi/model_creators/specific_creators.py
@@ -8,10 +8,10 @@ from typing import Iterable
 from typing import List
 from typing import Optional
 
+from pcapi.core.offerers.models import Offerer
 from pcapi.models import Criterion
 from pcapi.models import EventType
 from pcapi.models import Offer
-from pcapi.models import Offerer
 from pcapi.models import Product
 from pcapi.models import Provider
 from pcapi.models import Stock
