+++ src/pcapi/models/__init__.py
@@ -1,7 +1,7 @@
 from pcapi.core.bookings.models import Booking
 from pcapi.core.mails.models import Email
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.offerers.models import Venue
-from pcapi.core.offerers.offerer import Offerer
 from pcapi.core.offers.models import Mediation
 from pcapi.core.offers.models import Offer
 from pcapi.core.offers.models import Stock
