+++ src/pcapi/sandboxes/scripts/creators/helpers/sql_creators.py
@@ -7,11 +7,11 @@ from typing import Iterable
 from typing import Optional
 
 from pcapi.core.bookings import api as bookings_api
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.users.models import User
 from pcapi.models import Booking
 from pcapi.models import EventType
 from pcapi.models import Offer
-from pcapi.models import Offerer
 from pcapi.models import Payment
 from pcapi.models import PaymentMessage
 from pcapi.models import Product
