+++ src/pcapi/sandboxes/scripts/getters/pro_04_offerer.py
@@ -1,4 +1,4 @@
-from pcapi.core.offerers.offerer import Offerer
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.users.models import User
 from pcapi.models.user_offerer import UserOfferer
 from pcapi.sandboxes.scripts.utils.helpers import get_offerer_helper
