+++ src/pcapi/sandboxes/scripts/getters/pro_08_stocks.py
@@ -1,4 +1,4 @@
-from pcapi.core.offerers.offerer import Offerer
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.offers.models import Offer
 from pcapi.core.users.models import User
 from pcapi.models import BankInformation
