+++ src/pcapi/sandboxes/scripts/getters/pro_13_reimbursements.py
@@ -1,5 +1,5 @@
 from pcapi.core.bookings.models import Booking
-from pcapi.core.offerers.offerer import Offerer
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.offers.models import Offer
 from pcapi.core.offers.models import Stock
 from pcapi.core.users.models import User
