+++ src/pcapi/sandboxes/scripts/getters/webapp_08_booking.py
@@ -1,8 +1,8 @@
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.offers.models import Mediation
 from pcapi.core.users.models import User
 from pcapi.models import EventType
 from pcapi.models import Offer
-from pcapi.models import Offerer
 from pcapi.models import Stock
 from pcapi.models import ThingType
 from pcapi.models import Venue
