+++ src/pcapi/validation/models/entity_validator.py
@@ -1,9 +1,9 @@
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.users.models import User
 from pcapi.models import ApiErrors
 from pcapi.models import BankInformation
 from pcapi.models import HasAddressMixin
 from pcapi.models import Offer
-from pcapi.models import Offerer
 from pcapi.models import Product
 from pcapi.models import Stock
 from pcapi.models import Venue
