+++ src/pcapi/validation/models/offerer.py
@@ -1,5 +1,5 @@
+from pcapi.core.offerers.models import Offerer
 from pcapi.models import ApiErrors
-from pcapi.models import Offerer
 
 
 def validate(offerer: Offerer, api_errors: ApiErrors) -> ApiErrors:
