+++ tests/domain/payments_test.py
@@ -5,6 +5,7 @@ import uuid
 from freezegun import freeze_time
 import pytest
 
+from pcapi.core.offerers.models import Offerer
 from pcapi.domain.payments import UnmatchedPayments
 from pcapi.domain.payments import apply_banishment
 from pcapi.domain.payments import create_all_payments_details
@@ -28,7 +29,6 @@ from pcapi.model_creators.generic_creators import create_venue
 from pcapi.model_creators.specific_creators import create_offer_with_thing_product
 from pcapi.models import Booking
 from pcapi.models import Offer
-from pcapi.models import Offerer
 from pcapi.models import Venue
 from pcapi.models.payment import Payment
 from pcapi.models.payment_status import TransactionStatus
