+++ tests/routes/pro/post_offerer_test.py
@@ -4,11 +4,11 @@ from unittest.mock import patch
 
 import pytest
 
+from pcapi.core.offerers.models import Offerer
 from pcapi.model_creators.generic_creators import create_offerer
 from pcapi.model_creators.generic_creators import create_user
 from pcapi.model_creators.generic_creators import create_user_offerer
 from pcapi.model_creators.generic_creators import create_venue_type
-from pcapi.models import Offerer
 from pcapi.models import UserOfferer
 from pcapi.repository import repository
 from pcapi.utils.human_ids import humanize
