+++ tests/routes/pro/signup_pro_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pcapi.core.offerers.offerer import Offerer
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.users.models import User
 from pcapi.model_creators.generic_creators import create_user
 from pcapi.model_creators.generic_creators import create_user_offerer
