+++ tests/routes/pro/validate_user_test.py
@@ -2,11 +2,11 @@ from unittest.mock import patch
 
 import pytest
 
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.users.models import User
 from pcapi.model_creators.generic_creators import create_offerer
 from pcapi.model_creators.generic_creators import create_user
 from pcapi.model_creators.generic_creators import create_user_offerer
-from pcapi.models import Offerer
 from pcapi.models import UserOfferer
 from pcapi.repository import repository
