+++ tests/scripts/offerer/file_import_test.py
@@ -6,11 +6,11 @@ from unittest.mock import patch
 from freezegun import freeze_time
 import pytest
 
+from pcapi.core.offerers.models import Offerer
 from pcapi.core.users.models import User
 from pcapi.model_creators.generic_creators import create_offerer
 from pcapi.model_creators.generic_creators import create_user
 from pcapi.model_creators.generic_creators import create_venue_type
-from pcapi.models import Offerer
 from pcapi.models import UserOfferer
 from pcapi.repository import repository
 from pcapi.scripts.offerer.file_import import OffererNotCreatedException
